BalloonMessages can draw itself - Move to VisualObject
