+++ org.uqbar.project.wollok.game/src/org/uqbar/project/wollok/game/Balloon.xtend
@@ -1,19 +1,19 @@
 package org.uqbar.project.wollok.game
 
-import com.badlogic.gdx.graphics.g2d.NinePatch
-import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.Gdx
-import org.uqbar.project.wollok.game.gameboard.Gameboard
-import com.badlogic.gdx.graphics.g2d.SpriteBatch
+import com.badlogic.gdx.graphics.Color
+import com.badlogic.gdx.graphics.Texture
 import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.graphics.g2d.GlyphLayout
-import com.badlogic.gdx.graphics.Color
+import com.badlogic.gdx.graphics.g2d.NinePatch
+import com.badlogic.gdx.graphics.g2d.SpriteBatch
+import org.uqbar.project.wollok.game.gameboard.Gameboard
 
 class Balloon {
 
 	val static NinePatch patch = new NinePatch(new Texture(Gdx.files.internal("assets/speech.png")), 30, 60, 40, 50)
+	val static colorNegro = Color.RED // Color.BLACK
 	val width = 75
-	val static colorNegro = new Color(0,0,0,100)
 	val static textBitmap = new BitmapFont()
 	var plusWidth = 0
 	var glyphLayout = new GlyphLayout()
@@ -25,11 +25,12 @@ class Balloon {
 		
 		plusWidth = 0	
 		glyphLayout.reset
-		glyphLayout.setText(textBitmap,newText,colorNegro,width,3,true)
+		this.setText(newText, width)
+		
 		while(glyphLayout.height > 29){
 			glyphLayout.reset
 			plusWidth += 10
-			glyphLayout.setText(textBitmap,newText,colorNegro,width + plusWidth,3,true)
+			this.setText(newText, width + plusWidth)
 		}
 		
 		var character = Gameboard.getInstance.getCharacter()
@@ -38,5 +39,9 @@ class Balloon {
 		patch.draw(batch, x, y, 100 + plusWidth, 85)
 		
 		textBitmap.draw(batch, glyphLayout, x + 10, y + 75);
+	}
+	
+	private def setText(String text, int width) {
+		glyphLayout.setText(textBitmap, text, colorNegro, width, 3, true)
 	}
 }
