+++ org.uqbar.project.wollok.game/src/org/uqbar/project/wollok/game/gameboard/GameboardRendering.java
@@ -48,12 +48,16 @@ public class GameboardRendering implements ApplicationListener {
 		for (int i=0; i < gameboard.getListeners().size(); i++){
 			try {
 				gameboard.getListeners().get(i).notify(gameboard);
-			} catch (WollokProgramExceptionWrapper e) {
+			} 
+			catch (WollokProgramExceptionWrapper e) {
 				Object message = e.getWollokException().getInstanceVariables().get("message");
 				if (message == null)
-					message = "!!!";
-				gameboard.characterSay(message.toString());
-			} catch (WollokInterpreterException e) { } // TODO 
+					message = "NO MESSAGE";
+				gameboard.characterSay("ERROR: " + message.toString());
+			} 
+//			catch (WollokInterpreterException e) { 
+//				gameboard.characterSay(e.getCause().getCause().getMessage()); 
+//			}
 
 		}
 
@@ -65,9 +69,6 @@ public class GameboardRendering implements ApplicationListener {
 			this.draw(component);
 		}
 
-		if (gameboard.hasMessages())
-			balloon.draw(batch,gameboard.getCurrentMessage());
-		
 		batch.end();
 	}
